Refactor code around setting sequences while switching primary traffic
