+++ go/vt/vtctl/workflow/traffic_switcher_test.go
@@ -27,19 +27,14 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"vitess.io/vitess/go/sqlescape"
 	"vitess.io/vitess/go/sqltypes"
-	"vitess.io/vitess/go/vt/mysqlctl/tmutils"
 	"vitess.io/vitess/go/vt/proto/binlogdata"
+	tabletmanagerdatapb "vitess.io/vitess/go/vt/proto/tabletmanagerdata"
+	topodatapb "vitess.io/vitess/go/vt/proto/topodata"
 	"vitess.io/vitess/go/vt/proto/vschema"
 	"vitess.io/vitess/go/vt/sqlparser"
 	"vitess.io/vitess/go/vt/topo"
 	"vitess.io/vitess/go/vt/vtgate/vindexes"
-	"vitess.io/vitess/go/vt/vttablet/tabletmanager/vreplication"
-
-	tabletmanagerdatapb "vitess.io/vitess/go/vt/proto/tabletmanagerdata"
-	topodatapb "vitess.io/vitess/go/vt/proto/topodata"
-	vtctldatapb "vitess.io/vitess/go/vt/proto/vtctldata"
 )
 
 type testTrafficSwitcher struct {
@@ -72,521 +67,6 @@ func TestReverseWorkflowName(t *testing.T) {
 	for _, test := range tests {
 		got := ReverseWorkflowName(test.in)
 		assert.Equal(t, test.out, got)
-	}
-}
-
-func TestGetTargetSequenceMetadata(t *testing.T) {
-	ctx, cancel := context.WithTimeout(context.Background(), 60*time.Second)
-	defer cancel()
-	cell := "cell1"
-	workflow := "wf1"
-	table := "`t1`"
-	tableDDL := "create table t1 (id int not null auto_increment primary key, c1 varchar(10))"
-	table2 := "t2"
-	unescapedTable := "t1"
-	sourceKeyspace := &testKeyspace{
-		KeyspaceName: "source-ks",
-		ShardNames:   []string{"0"},
-	}
-	targetKeyspace := &testKeyspace{
-		KeyspaceName: "targetks",
-		ShardNames:   []string{"-80", "80-"},
-	}
-	vindexes := map[string]*vschema.Vindex{
-		"xxhash": {
-			Type: "xxhash",
-		},
-	}
-	env := newTestEnv(t, ctx, cell, sourceKeyspace, targetKeyspace)
-	defer env.close()
-
-	env.tmc.schema = map[string]*tabletmanagerdatapb.SchemaDefinition{
-		unescapedTable: {
-			TableDefinitions: []*tabletmanagerdatapb.TableDefinition{
-				{
-					Name:   unescapedTable,
-					Schema: tableDDL,
-				},
-			},
-		},
-	}
-
-	type testCase struct {
-		name                                   string
-		sourceVSchema                          *vschema.Keyspace
-		targetVSchema                          *vschema.Keyspace
-		options                                *vtctldatapb.WorkflowOptions
-		want                                   map[string]*sequenceMetadata
-		expectSourceApplySchemaRequestResponse *applySchemaRequestResponse
-		err                                    string
-	}
-	tests := []testCase{
-		{
-			name: "no sequences",
-			want: nil,
-		},
-		{
-			name: "sequences with backticks and qualified table",
-			sourceVSchema: &vschema.Keyspace{
-				Vindexes: vindexes,
-				Tables: map[string]*vschema.Table{
-					"`my-seq1`": {
-						Type: "sequence",
-					},
-				},
-			},
-			targetVSchema: &vschema.Keyspace{
-				Vindexes: vindexes,
-				Tables: map[string]*vschema.Table{
-					table: {
-						ColumnVindexes: []*vschema.ColumnVindex{
-							{
-								Name:   "xxhash",
-								Column: "`my-col`",
-							},
-						},
-						AutoIncrement: &vschema.AutoIncrement{
-							Column:   "`my-col`",
-							Sequence: fmt.Sprintf("`%s`.`my-seq1`", sourceKeyspace.KeyspaceName),
-						},
-					},
-				},
-			},
-			want: map[string]*sequenceMetadata{
-				"my-seq1": {
-					backingTableName:     "my-seq1",
-					backingTableKeyspace: "source-ks",
-					backingTableDBName:   "vt_source-ks",
-					usingTableName:       unescapedTable,
-					usingTableDBName:     "vt_targetks",
-					usingTableDefinition: &vschema.Table{
-						ColumnVindexes: []*vschema.ColumnVindex{
-							{
-								Column: "my-col",
-								Name:   "xxhash",
-							},
-						},
-						AutoIncrement: &vschema.AutoIncrement{
-							Column:   "my-col",
-							Sequence: fmt.Sprintf("%s.my-seq1", sourceKeyspace.KeyspaceName),
-						},
-					},
-				},
-			},
-		},
-		{
-			name: "auto_increment replaced with sequence",
-			sourceVSchema: &vschema.Keyspace{
-				Vindexes: vindexes,
-				Tables:   map[string]*vschema.Table{}, // Sequence table will be created
-			},
-			options: &vtctldatapb.WorkflowOptions{
-				ShardedAutoIncrementHandling: vtctldatapb.ShardedAutoIncrementHandling_REPLACE,
-				GlobalKeyspace:               sourceKeyspace.KeyspaceName,
-			},
-			expectSourceApplySchemaRequestResponse: &applySchemaRequestResponse{
-				change: &tmutils.SchemaChange{
-					SQL: sqlparser.BuildParsedQuery(sqlCreateSequenceTable,
-						sqlescape.EscapeID(fmt.Sprintf(autoSequenceTableFormat, unescapedTable))).Query,
-					Force:                   false,
-					AllowReplication:        true,
-					SQLMode:                 vreplication.SQLMode,
-					DisableForeignKeyChecks: true,
-				},
-				res: &tabletmanagerdatapb.SchemaChangeResult{},
-			},
-			targetVSchema: &vschema.Keyspace{
-				Vindexes: vindexes,
-				Tables: map[string]*vschema.Table{
-					table: {
-						ColumnVindexes: []*vschema.ColumnVindex{
-							{
-								Name:   "xxhash",
-								Column: "`my-col`",
-							},
-						},
-						AutoIncrement: &vschema.AutoIncrement{
-							Column:   "my-col",
-							Sequence: fmt.Sprintf(autoSequenceTableFormat, unescapedTable),
-						},
-					},
-				},
-			},
-			want: map[string]*sequenceMetadata{
-				fmt.Sprintf(autoSequenceTableFormat, unescapedTable): {
-					backingTableName:     fmt.Sprintf(autoSequenceTableFormat, unescapedTable),
-					backingTableKeyspace: "source-ks",
-					backingTableDBName:   "vt_source-ks",
-					usingTableName:       unescapedTable,
-					usingTableDBName:     "vt_targetks",
-					usingTableDefinition: &vschema.Table{
-						ColumnVindexes: []*vschema.ColumnVindex{
-							{
-								Column: "my-col",
-								Name:   "xxhash",
-							},
-						},
-						AutoIncrement: &vschema.AutoIncrement{
-							Column:   "my-col",
-							Sequence: fmt.Sprintf(autoSequenceTableFormat, unescapedTable),
-						},
-					},
-				},
-			},
-		},
-		{
-			name: "sequences with backticks",
-			sourceVSchema: &vschema.Keyspace{
-				Vindexes: vindexes,
-				Tables: map[string]*vschema.Table{
-					"`my-seq1`": {
-						Type: "sequence",
-					},
-				},
-			},
-			targetVSchema: &vschema.Keyspace{
-				Vindexes: vindexes,
-				Tables: map[string]*vschema.Table{
-					table: {
-						ColumnVindexes: []*vschema.ColumnVindex{
-							{
-								Name:   "xxhash",
-								Column: "`my-col`",
-							},
-						},
-						AutoIncrement: &vschema.AutoIncrement{
-							Column:   "`my-col`",
-							Sequence: "`my-seq1`",
-						},
-					},
-				},
-			},
-			want: map[string]*sequenceMetadata{
-				"my-seq1": {
-					backingTableName:     "my-seq1",
-					backingTableKeyspace: "source-ks",
-					backingTableDBName:   "vt_source-ks",
-					usingTableName:       unescapedTable,
-					usingTableDBName:     "vt_targetks",
-					usingTableDefinition: &vschema.Table{
-						ColumnVindexes: []*vschema.ColumnVindex{
-							{
-								Column: "my-col",
-								Name:   "xxhash",
-							},
-						},
-						AutoIncrement: &vschema.AutoIncrement{
-							Column:   "my-col",
-							Sequence: "my-seq1",
-						},
-					},
-				},
-			},
-		},
-		{
-			name: "sequences using vindexes with both column definition structures",
-			sourceVSchema: &vschema.Keyspace{
-				Vindexes: vindexes,
-				Tables: map[string]*vschema.Table{
-					"seq1": {
-						Type: "sequence",
-					},
-					"seq2": {
-						Type: "sequence",
-					},
-				},
-			},
-			targetVSchema: &vschema.Keyspace{
-				Vindexes: vindexes,
-				Tables: map[string]*vschema.Table{
-					table: {
-						ColumnVindexes: []*vschema.ColumnVindex{
-							{
-								Name:   "xxhash",
-								Column: "col1",
-							},
-						},
-						AutoIncrement: &vschema.AutoIncrement{
-							Column:   "col1",
-							Sequence: fmt.Sprintf("%s.seq1", sourceKeyspace.KeyspaceName),
-						},
-					},
-					table2: {
-						ColumnVindexes: []*vschema.ColumnVindex{
-							{
-								Name:    "xxhash",
-								Columns: []string{"col2"},
-							},
-						},
-						AutoIncrement: &vschema.AutoIncrement{
-							Column:   "col2",
-							Sequence: fmt.Sprintf("%s.seq2", sourceKeyspace.KeyspaceName),
-						},
-					},
-				},
-			},
-			want: map[string]*sequenceMetadata{
-				"seq1": {
-					backingTableName:     "seq1",
-					backingTableKeyspace: "source-ks",
-					backingTableDBName:   "vt_source-ks",
-					usingTableName:       unescapedTable,
-					usingTableDBName:     "vt_targetks",
-					usingTableDefinition: &vschema.Table{
-						ColumnVindexes: []*vschema.ColumnVindex{
-							{
-								Column: "col1",
-								Name:   "xxhash",
-							},
-						},
-						AutoIncrement: &vschema.AutoIncrement{
-							Column:   "col1",
-							Sequence: fmt.Sprintf("%s.seq1", sourceKeyspace.KeyspaceName),
-						},
-					},
-				},
-				"seq2": {
-					backingTableName:     "seq2",
-					backingTableKeyspace: "source-ks",
-					backingTableDBName:   "vt_source-ks",
-					usingTableName:       table2,
-					usingTableDBName:     "vt_targetks",
-					usingTableDefinition: &vschema.Table{
-						ColumnVindexes: []*vschema.ColumnVindex{
-							{
-								Columns: []string{"col2"},
-								Name:    "xxhash",
-							},
-						},
-						AutoIncrement: &vschema.AutoIncrement{
-							Column:   "col2",
-							Sequence: fmt.Sprintf("%s.seq2", sourceKeyspace.KeyspaceName),
-						},
-					},
-				},
-			},
-		},
-		{
-			name: "sequence with table having mult-col vindex",
-			sourceVSchema: &vschema.Keyspace{
-				Vindexes: vindexes,
-				Tables: map[string]*vschema.Table{
-					"seq1": {
-						Type: "sequence",
-					},
-				},
-			},
-			targetVSchema: &vschema.Keyspace{
-				Vindexes: vindexes,
-				Tables: map[string]*vschema.Table{
-					table: {
-						ColumnVindexes: []*vschema.ColumnVindex{
-							{
-								Name:    "xxhash",
-								Columns: []string{"col3", "col4"},
-							},
-						},
-						AutoIncrement: &vschema.AutoIncrement{
-							Column:   "col1",
-							Sequence: fmt.Sprintf("%s.seq1", sourceKeyspace.KeyspaceName),
-						},
-					},
-				},
-			},
-			want: map[string]*sequenceMetadata{
-				"seq1": {
-					backingTableName:     "seq1",
-					backingTableKeyspace: "source-ks",
-					backingTableDBName:   "vt_source-ks",
-					usingTableName:       unescapedTable,
-					usingTableDBName:     "vt_targetks",
-					usingTableDefinition: &vschema.Table{
-						ColumnVindexes: []*vschema.ColumnVindex{
-							{
-								Columns: []string{"col3", "col4"},
-								Name:    "xxhash",
-							},
-						},
-						AutoIncrement: &vschema.AutoIncrement{
-							Column:   "col1",
-							Sequence: fmt.Sprintf("%s.seq1", sourceKeyspace.KeyspaceName),
-						},
-					},
-				},
-			},
-		},
-		{
-			name: "invalid table name",
-			sourceVSchema: &vschema.Keyspace{
-				Vindexes: vindexes,
-				Tables: map[string]*vschema.Table{
-					"`my-`seq1`": {
-						Type: "sequence",
-					},
-				},
-			},
-			targetVSchema: &vschema.Keyspace{
-				Vindexes: vindexes,
-				Tables: map[string]*vschema.Table{
-					table: {
-						ColumnVindexes: []*vschema.ColumnVindex{
-							{
-								Name:   "xxhash",
-								Column: "`my-col`",
-							},
-						},
-						AutoIncrement: &vschema.AutoIncrement{
-							Column:   "`my-col`",
-							Sequence: "`my-seq1`",
-						},
-					},
-				},
-			},
-			err: "invalid table name \"`my-`seq1`\" in keyspace source-ks: UnescapeID err: unexpected single backtick at position 3 in 'my-`seq1'",
-		},
-		{
-			name: "invalid keyspace name",
-			sourceVSchema: &vschema.Keyspace{
-				Vindexes: vindexes,
-				Tables: map[string]*vschema.Table{
-					"`my-seq1`": {
-						Type: "sequence",
-					},
-				},
-			},
-			targetVSchema: &vschema.Keyspace{
-				Vindexes: vindexes,
-				Tables: map[string]*vschema.Table{
-					table: {
-						ColumnVindexes: []*vschema.ColumnVindex{
-							{
-								Name:   "xxhash",
-								Column: "`my-col`",
-							},
-						},
-						AutoIncrement: &vschema.AutoIncrement{
-							Column:   "`my-col`",
-							Sequence: "`ks`1`.`my-seq1`",
-						},
-					},
-				},
-			},
-			err: "invalid keyspace in qualified sequence table name \"`ks`1`.`my-seq1`\" defined in sequence table column_vindexes:{column:\"`my-col`\" name:\"xxhash\"} auto_increment:{column:\"`my-col`\" sequence:\"`ks`1`.`my-seq1`\"}: UnescapeID err: unexpected single backtick at position 2 in 'ks`1'",
-		},
-		{
-			name: "invalid auto-inc column name",
-			sourceVSchema: &vschema.Keyspace{
-				Vindexes: vindexes,
-				Tables: map[string]*vschema.Table{
-					"`my-seq1`": {
-						Type: "sequence",
-					},
-				},
-			},
-			targetVSchema: &vschema.Keyspace{
-				Vindexes: vindexes,
-				Tables: map[string]*vschema.Table{
-					table: {
-						ColumnVindexes: []*vschema.ColumnVindex{
-							{
-								Name:   "xxhash",
-								Column: "`my-col`",
-							},
-						},
-						AutoIncrement: &vschema.AutoIncrement{
-							Column:   "`my`-col`",
-							Sequence: "`my-seq1`",
-						},
-					},
-				},
-			},
-			err: "invalid auto-increment column name \"`my`-col`\" defined in sequence table column_vindexes:{column:\"my-col\" name:\"xxhash\"} auto_increment:{column:\"`my`-col`\" sequence:\"my-seq1\"}: UnescapeID err: unexpected single backtick at position 2 in 'my`-col'",
-		},
-		{
-			name: "invalid sequence name",
-			sourceVSchema: &vschema.Keyspace{
-				Vindexes: vindexes,
-				Tables: map[string]*vschema.Table{
-					"`my-seq1`": {
-						Type: "sequence",
-					},
-				},
-			},
-			targetVSchema: &vschema.Keyspace{
-				Vindexes: vindexes,
-				Tables: map[string]*vschema.Table{
-					table: {
-						ColumnVindexes: []*vschema.ColumnVindex{
-							{
-								Name:   "xxhash",
-								Column: "`my-col`",
-							},
-						},
-						AutoIncrement: &vschema.AutoIncrement{
-							Column:   "`my-col`",
-							Sequence: "`my-`seq1`",
-						},
-					},
-				},
-			},
-			err: "invalid sequence table name \"`my-`seq1`\" defined in sequence table column_vindexes:{column:\"`my-col`\" name:\"xxhash\"} auto_increment:{column:\"`my-col`\" sequence:\"`my-`seq1`\"}: UnescapeID err: unexpected single backtick at position 3 in 'my-`seq1'",
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			err := env.ts.SaveVSchema(ctx, &topo.KeyspaceVSchemaInfo{
-				Name:     sourceKeyspace.KeyspaceName,
-				Keyspace: tc.sourceVSchema,
-			})
-			require.NoError(t, err)
-			err = env.ts.SaveVSchema(ctx, &topo.KeyspaceVSchemaInfo{
-				Name:     targetKeyspace.KeyspaceName,
-				Keyspace: tc.targetVSchema,
-			})
-			require.NoError(t, err)
-			err = env.ts.RebuildSrvVSchema(ctx, nil)
-			require.NoError(t, err)
-			sources := make(map[string]*MigrationSource, len(sourceKeyspace.ShardNames))
-			targets := make(map[string]*MigrationTarget, len(targetKeyspace.ShardNames))
-			for i, shard := range sourceKeyspace.ShardNames {
-				tablet := env.tablets[sourceKeyspace.KeyspaceName][startingSourceTabletUID+(i*tabletUIDStep)]
-				sources[shard] = &MigrationSource{
-					primary: &topo.TabletInfo{
-						Tablet: tablet,
-					},
-				}
-				if tc.expectSourceApplySchemaRequestResponse != nil {
-					env.tmc.expectApplySchemaRequest(tablet.Alias.Uid, tc.expectSourceApplySchemaRequestResponse)
-				}
-			}
-			for i, shard := range targetKeyspace.ShardNames {
-				tablet := env.tablets[targetKeyspace.KeyspaceName][startingTargetTabletUID+(i*tabletUIDStep)]
-				targets[shard] = &MigrationTarget{
-					primary: &topo.TabletInfo{
-						Tablet: tablet,
-					},
-				}
-			}
-			ts := &trafficSwitcher{
-				id:             1,
-				ws:             env.ws,
-				workflow:       workflow,
-				tables:         []string{table, table2},
-				sourceKeyspace: sourceKeyspace.KeyspaceName,
-				targetKeyspace: targetKeyspace.KeyspaceName,
-				sources:        sources,
-				targets:        targets,
-				options:        tc.options,
-			}
-			got, err := ts.getTargetSequenceMetadata(ctx)
-			if tc.err != "" {
-				require.EqualError(t, err, tc.err)
-			} else {
-				require.NoError(t, err)
-			}
-			require.EqualValues(t, tc.want, got)
-		})
 	}
 }
 
@@ -657,71 +137,6 @@ func TestTrafficSwitchPositionHandling(t *testing.T) {
 		return nil
 	})
 	require.NoError(t, err)
-}
-
-func TestInitializeTargetSequences(t *testing.T) {
-	ctx, cancel := context.WithTimeout(context.Background(), 60*time.Second)
-	defer cancel()
-
-	workflowName := "wf1"
-	tableName := "t1"
-	sourceKeyspaceName := "sourceks"
-	targetKeyspaceName := "targetks"
-
-	schema := map[string]*tabletmanagerdatapb.SchemaDefinition{
-		tableName: {
-			TableDefinitions: []*tabletmanagerdatapb.TableDefinition{
-				{
-					Name:   tableName,
-					Schema: fmt.Sprintf("CREATE TABLE %s (id BIGINT, name VARCHAR(64), PRIMARY KEY (id))", tableName),
-				},
-			},
-		},
-	}
-
-	sourceKeyspace := &testKeyspace{
-		KeyspaceName: sourceKeyspaceName,
-		ShardNames:   []string{"0"},
-	}
-	targetKeyspace := &testKeyspace{
-		KeyspaceName: targetKeyspaceName,
-		ShardNames:   []string{"0"},
-	}
-
-	env := newTestEnv(t, ctx, defaultCellName, sourceKeyspace, targetKeyspace)
-	defer env.close()
-	env.tmc.schema = schema
-
-	ts, _, err := env.ws.getWorkflowState(ctx, targetKeyspaceName, workflowName)
-	require.NoError(t, err)
-	sw := &switcher{ts: ts, s: env.ws}
-
-	sequencesByBackingTable := map[string]*sequenceMetadata{
-		"my-seq1": {
-			backingTableName:     "my-seq1",
-			backingTableKeyspace: sourceKeyspaceName,
-			backingTableDBName:   fmt.Sprintf("vt_%s", sourceKeyspaceName),
-			usingTableName:       tableName,
-			usingTableDBName:     "vt_targetks",
-			usingTableDefinition: &vschema.Table{
-				AutoIncrement: &vschema.AutoIncrement{
-					Column:   "my-col",
-					Sequence: fmt.Sprintf("%s.my-seq1", sourceKeyspace.KeyspaceName),
-				},
-			},
-		},
-	}
-
-	env.tmc.expectVRQuery(200, "/select max.*", sqltypes.MakeTestResult(sqltypes.MakeTestFields("maxval", "int64"), "34"))
-	// Expect the insert query to be executed with 35 as a params, since we provide a maxID of 34 in the last query
-	env.tmc.expectVRQuery(100, "/insert into.*35.*", &sqltypes.Result{RowsAffected: 1})
-
-	err = sw.initializeTargetSequences(ctx, sequencesByBackingTable)
-	assert.NoError(t, err)
-
-	// Expect the queries to be cleared
-	assert.Empty(t, env.tmc.vrQueries[100])
-	assert.Empty(t, env.tmc.vrQueries[200])
 }
 
 func TestAddTenantFilter(t *testing.T) {
